tg-bot/requirements.txt
python-telegram-bot~=21.0.1
httpx~=0.27
python-dotenv~=1.0.0
a2a-sdk~=0.3.10